+++ gDRplots/R/plot-sa-response_curve.R
@@ -8,8 +8,8 @@
 #'    (that is when rownames(se) == 1 it has to be "cId", otherwise - "rId")
 #' @param group_names character vector with names to subset from se (the same dim as \code{grouping});
 #'    if \code{NULL} then all values will be plotted
-#' @param normalization_type string with normalization_types to be selected
-#'                           one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
+#' @param metric_growth string with normalization_types to be selected
+#'                      one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
 #' @param colormap character vector with colors for \code{group_names} - name or hex value
 #' @param plot_averaged_flag logical flag whether plot points with average values 
 #' @param plot_fit_flag logical flag whether plot points with fitted values 
@@ -19,57 +19,66 @@
 #' @keywords single-agent_plots
 #' @examples
 #' mae <- gDRutils::get_synthetic_data("small")
-#' se <- mae[[1]]
-#' iR <-  rownames(se)[1]
-#' grouping <- "cId"
+#' se <- mae[[gDRutils::get_supported_experiments("sa")]]
+#' iR <- rownames(se)[1]
+#' grouping <- "CellLineName"
 #' dt_metrics <- gDRutils::convert_se_assay_to_dt(se[iR], "Metrics")
 #' dt_average <- gDRutils::convert_se_assay_to_dt(se[iR], "Averaged")
-#' group_names <- colnames(se)[2:5]
+#' group_names <- unique(dt_metrics[[grouping]])[1:3]
 #' 
-#' grob_sa(dt_metrics = dt_metrics, 
-#'         dt_average = dt_average, 
-#'         grouping = grouping,
-#'         group_names = group_names)
+#' plot_dose_response_sa(dt_metrics = dt_metrics, 
+#'                       dt_average = dt_average, 
+#'                       grouping = grouping,
+#'                       group_names = group_names)
+#'             
+#' iC <- colnames(se)[1]                    
+#' grouping <- "DrugName"
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se[, iC], "Metrics")
+#' dt_average <- gDRutils::convert_se_assay_to_dt(se[, iC], "Averaged")
+#' group_names <- unique(dt_metrics[[grouping]])[1:3]
 #' 
+#' plot_dose_response_sa(dt_metrics = dt_metrics, 
+#'                       dt_average = dt_average, 
+#'                       grouping = grouping,
+#'                       group_names = group_names)
+#'                       
 #' @export
-grob_sa <- function(dt_metrics, 
-                    dt_average, 
-                    grouping, 
-                    group_names = NULL, 
-                    normalization_type = "GR", 
-                    colormap = NULL, 
-                    plot_averaged_flag = TRUE, 
-                    plot_fit_flag = TRUE) {
-  
-  checkmate::expect_data_table(dt_metrics)
-  checkmate::expect_data_table(dt_average)
-  checkmate::expect_choice(grouping, choices = c("cId", "rId"))
-  checkmate::expect_character(group_names, null.ok = TRUE)
-  checkmate::expect_choice(normalization_type, choices = c("GR", "RV"))
-  checkmate::expect_character(colormap, null.ok = TRUE)
-  checkmate::expect_flag(plot_averaged_flag)
-  checkmate::expect_flag(plot_fit_flag)
+plot_dose_response_sa <- function(dt_metrics, 
+                                  dt_average, 
+                                  grouping, 
+                                  group_names = NULL, 
+                                  metric_growth = "GR", 
+                                  colormap = NULL, 
+                                  plot_averaged_flag = TRUE, 
+                                  plot_fit_flag = TRUE) {
   
   cellline_name <- gDRutils::get_env_identifiers("cellline_name")
   clid <- gDRutils::get_env_identifiers("cellline")
   drug_name <- gDRutils::get_env_identifiers("drug_name")
   gnumber <- gDRutils::get_env_identifiers("drug")
   
+  checkmate::expect_data_table(dt_metrics)
+  checkmate::expect_data_table(dt_average)
+  checkmate::expect_choice(grouping, choices = c(cellline_name, drug_name))
+  checkmate::expect_character(group_names, null.ok = TRUE)
+  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
+  checkmate::expect_character(colormap, null.ok = TRUE)
+  checkmate::expect_flag(plot_averaged_flag)
+  checkmate::expect_flag(plot_fit_flag)
+  
   # check input data
-  if (grouping == "cId") {
+  if (grouping == cellline_name) {
     stopifnot("grouping` does not fit to `dt_metrics` and `dt_average`" =
-                (NROW(unique(dt_metrics[["rId"]])) == 1 && NROW(unique(dt_average[["rId"]])) == 1))
-  } else if (grouping == "rId") {
+                (NROW(unique(dt_metrics[[drug_name]])) == 1 && NROW(unique(dt_average[[drug_name]])) == 1))
+  } else if (grouping == drug_name) {
     stopifnot("grouping` does not fit to `dt_metrics` and `dt_average`" =
-                (NROW(unique(dt_metrics[["cId"]])) == 1 && NROW(unique(dt_average[["cId"]])) == 1))
+                (NROW(unique(dt_metrics[[cellline_name]])) == 1 && NROW(unique(dt_average[[cellline_name]])) == 1))
   }
   stopifnot("empty plot was selected" = any(plot_averaged_flag, plot_fit_flag))
   
   # filter data for normalization type
-  data.table::setkeyv(dt_metrics, "normalization_type")
-  dt_met_norm <- dt_metrics[normalization_type]
-  data.table::setkeyv(dt_average, "normalization_type")
-  dt_avg_norm <- dt_average[normalization_type]
+  dt_met_norm <- dt_metrics[normalization_type == metric_growth, ]
+  dt_avg_norm <- dt_average[normalization_type == metric_growth, ]
   
   # variables
   conc <- gDRutils::get_env_identifiers("concentration")
@@ -90,7 +99,7 @@ grob_sa <- function(dt_metrics,
   dt_fit <- data.table::data.table()
   
   for (icol in group_names) {
-    sel_metrics <- dt_met_norm[dt_met_norm[[grouping]] == icol, ]
+    sel_metrics <- dt_met_norm[get(grouping) == icol, ]
     dt_fit <- rbind(dt_fit, 
                     cbind(sel_metrics[, grouping, with = FALSE],
                           data.table::data.table(
@@ -125,11 +134,9 @@ grob_sa <- function(dt_metrics,
   dt_fit$grouping <- factor(dt_fit[[grouping]], levels = group_names)
   
   plt_title <- sprintf(
-    "%s Drug dose response for %s: \n%s (%s)",
-    normalization_type,
-    ifelse(grouping == "cId", "Drug", "Cell Line"),
-    ifelse(grouping == "cId", unique(dt_metrics[[drug_name]]), unique(dt_metrics[[cellline_name]])),
-    ifelse(grouping == "cId", unique(dt_metrics[[gnumber]]), unique(dt_metrics[[clid]]))
+    "%s (%s)",
+    ifelse(grouping == cellline_name, unique(dt_metrics[[drug_name]]), unique(dt_metrics[[cellline_name]])),
+    ifelse(grouping == cellline_name, unique(dt_metrics[[gnumber]]), unique(dt_metrics[[clid]]))
   )
   
   # final plot
@@ -142,7 +149,7 @@ grob_sa <- function(dt_metrics,
     ggplot2::coord_cartesian(xlim = conc_range, ylim = data_range) +
     ggplot2::scale_x_continuous(breaks = -5:2, labels = c("1e-5", "1e-4", 10 ^ (-3:2))) +
     ggplot2::xlab(bquote(.(conc) ~ "[" ~ mu * M ~ "]")) +
-    ggplot2::ylab(paste(normalization_type, "values")) +
+    ggplot2::ylab(metric_growth) +
     ggplot2::ggtitle(plt_title) +
     ggplot2::theme_bw()
   
@@ -164,119 +171,380 @@ grob_sa <- function(dt_metrics,
 
 #' Plot drug response curves for single-agent data for selected call lines and drugs
 #' 
-#' @inheritParams grob_sa
-#' @param se single-agent \code{SummarizedExperiment} object holding raw and/or processed 
-#'    dose-response data in its assays for one cell line
-#' @param cellline_name character vector with cell line to be plotted (colnames of se)
-#' @param drug_name character vector with cell line to be plotted (rownames of se)
+#' @inheritParams plot_dose_response_sa
+#' @param cellline_name_vec character vector with cell line to be plotted (Cell Line Name)
+#' @param drug_name_vec character vector with cell line to be plotted (Drug Name)
 #'    
 #' @return list of plots with dose-response curves
 #' 
 #' @keywords single-agent_plots
 #' @examples
 #' mae <- gDRutils::get_synthetic_data("small")
-#' se <- mae[[1]]
-#' cellline_name <- colnames(se)[2:5]
-#' drug_name <- rownames(se)[5:7]
+#' se <- mae[[gDRutils::get_supported_experiments("sa")]]
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
+#' dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
+#' cellline_name_vec <- unique(dt_metrics[["CellLineName"]])[2:5]
+#' drug_name_vec <- unique(dt_metrics[["DrugName"]])[5:7]
 #' 
-#' plot_sa_by_CLs(se = se, 
-#'                cellline_name = cellline_name, 
-#'                drug_name = drug_name, 
-#'                normalization_type = "RV", 
-#'                colormap = c("#B9D3EE", "#FF6347", "#C2F970"))
+#' plot_dose_response_sa_by_CLs(dt_metrics = dt_metrics, 
+#'                              dt_average = dt_average, 
+#'                              cellline_name_vec = cellline_name_vec, 
+#'                              drug_name_vec = drug_name_vec, 
+#'                              metric_growth = "RV", 
+#'                              colormap = c("#B9D3EE", "#FF6347", "#C2F970"))
 #' 
 #' @export
-plot_sa_by_CLs <- function(se, 
-                           cellline_name = NULL, 
-                           drug_name = NULL,
-                           normalization_type = "GR", 
-                           colormap = NULL, 
-                           plot_averaged_flag = TRUE, 
-                           plot_fit_flag = TRUE) {
+plot_dose_response_sa_by_CLs <- function(dt_metrics, 
+                                         dt_average,
+                                         cellline_name_vec = NULL, 
+                                         drug_name_vec = NULL,
+                                         metric_growth = "GR", 
+                                         colormap = NULL, 
+                                         plot_averaged_flag = TRUE, 
+                                         plot_fit_flag = TRUE) {
   
-  checkmate::assert_class(se, "SummarizedExperiment")
-  checkmate::expect_character(cellline_name, null.ok = TRUE)
-  checkmate::expect_character(drug_name, null.ok = TRUE)
-  checkmate::expect_choice(normalization_type, choices = c("GR", "RV"))
+  checkmate::expect_data_table(dt_metrics)
+  checkmate::expect_data_table(dt_average)
+  checkmate::expect_character(cellline_name_vec, null.ok = TRUE)
+  checkmate::expect_character(drug_name_vec, null.ok = TRUE)
+  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
   checkmate::expect_character(colormap, null.ok = TRUE)
   checkmate::expect_flag(plot_averaged_flag)
   checkmate::expect_flag(plot_fit_flag)
   
-  if (is.null(drug_name) || all(!drug_name %in% rownames(se))) {
-    drug_name  <- rownames(se)
-  } else if (!all(drug_name  %in% rownames(se))) {
-    drug_name <- drug_name[drug_name  %in% rownames(se)]
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  clid <- gDRutils::get_env_identifiers("cellline")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  gnumber <- gDRutils::get_env_identifiers("drug")
+  
+  available_drugs <- unique(dt_metrics[[drug_name]])
+  if (is.null(drug_name_vec) || all(!drug_name_vec %in% available_drugs)) {
+    drug_name_vec  <- available_drugs
+  } else if (!all(drug_name_vec %in% available_drugs)) {
+    drug_name_vec <- drug_name_vec[drug_name_vec  %in% available_drugs]
   }  
   
+  available_cellline <- unique(dt_metrics[[cellline_name]])
+  if (is.null(cellline_name_vec) || all(!cellline_name_vec %in% available_cellline)) {
+    cellline_name_vec <- available_cellline
+  } else if (!all(cellline_name_vec %in% available_cellline)) {
+    cellline_name_vec <- cellline_name_vec[cellline_name_vec  %in% available_cellline]
+  } 
+  
   plt_list <- list()
-  for (iR in drug_name) {
+  for (iR in drug_name_vec) {
     
-    if (is.null(cellline_name) || all(!cellline_name %in% colnames(se))) {
-      cellline_name <- colnames(se)
-    } else if (!all(cellline_name %in% colnames(se))) {
-      cellline_name <- cellline_name[cellline_name  %in% colnames(se)]
-    }  
+    # subset data
+    dt_metrics_subset <- dt_metrics[get(drug_name) == iR & get(cellline_name) %in% cellline_name_vec]
+    dt_average_subset <- dt_average[get(drug_name) == iR & get(cellline_name) %in% cellline_name_vec]
     
-    subset_se <- se[iR, cellline_name]
-    
-    plt_title <- paste(normalization_type, iR)
+    plt_title <- sprintf("%s (%s)", iR, unique(dt_metrics[get(drug_name) == iR, ][[gnumber]]))
     
     plt <- 
-      grob_sa(dt_metrics = gDRutils::convert_se_assay_to_dt(subset_se, "Metrics"), 
-              dt_average = gDRutils::convert_se_assay_to_dt(subset_se, "Averaged"), 
-              grouping = "cId",
-              group_names = cellline_name,
-              normalization_type = normalization_type,
-              colormap = colormap,
-              plot_averaged_flag = plot_averaged_flag,
-              plot_fit_flag = plot_fit_flag)
+      plot_dose_response_sa(dt_metrics = dt_metrics_subset, 
+                            dt_average = dt_average_subset, 
+                            grouping = cellline_name,
+                            group_names = cellline_name_vec,
+                            metric_growth = metric_growth,
+                            colormap = colormap,
+                            plot_averaged_flag = plot_averaged_flag,
+                            plot_fit_flag = plot_fit_flag)
     
     plt_list[[plt_title]] <- plt
-    
   }
   
   return(plt_list)
 }
 
-#' Plot drug response curves for single-agent data for one selected cell line
+#' Plot drug response curves for single-agent data for selected call lines and drugs
 #' 
-#' @inheritParams grob_sa
-#' @param se single-agent \code{SummarizedExperiment} object holding raw and/or processed 
-#'    dose-response data in its assays for one cell line
+#' @inheritParams plot_dose_response_sa
+#' @param cellline_name_vec character vector with cell line to be plotted (Cell Line Name)
+#' @param drug_name_vec character vector with cell line to be plotted (Drug Name)
 #'    
-#' @return plot with dose-response curves
+#' @return list of plots with dose-response curves
 #' 
+#' @keywords single-agent_plots
 #' @examples
-#' \dontrun{
 #' mae <- gDRutils::get_synthetic_data("small")
-#' se <- mae[[1]]
+#' se <- mae[[gDRutils::get_supported_experiments("sa")]]
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
+#' dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
+#' cellline_name_vec <- unique(dt_metrics[["CellLineName"]])[2:5]
+#' drug_name_vec <- unique(dt_metrics[["DrugName"]])[5:7]
 #' 
-#' plot_sa_1CL(se = se[,colnames(se)[1]], colormap = c("cadetblue", "orange", "darkblue"))
-#' }
+#' plot_dose_response_sa_by_drugs(dt_metrics = dt_metrics, 
+#'                                dt_average = dt_average, 
+#'                                cellline_name_vec = cellline_name_vec, 
+#'                                drug_name_vec = drug_name_vec, 
+#'                                metric_growth = "RV", 
+#'                                colormap = c("#B9D3EE", "#FF6347", "#C2F970"))
 #' 
-#' @keywords internal
-plot_sa_1CL <- function(se, 
-                        normalization_type = "GR", 
-                        colormap = NULL, 
-                        plot_averaged_flag = TRUE, 
-                        plot_fit_flag = TRUE) {
+#' @export
+plot_dose_response_sa_by_drugs <- function(dt_metrics, 
+                                           dt_average,
+                                           cellline_name_vec = NULL, 
+                                           drug_name_vec = NULL,
+                                           metric_growth = "GR", 
+                                           colormap = NULL, 
+                                           plot_averaged_flag = TRUE, 
+                                           plot_fit_flag = TRUE) {
   
-  stopifnot(NCOL(se) == 1) # plot for 1 cell line
-  
-  checkmate::assert_class(se, "SummarizedExperiment")
-  checkmate::expect_choice(normalization_type, choices = c("GR", "RV"))
+  checkmate::expect_data_table(dt_metrics)
+  checkmate::expect_data_table(dt_average)
+  checkmate::expect_character(cellline_name_vec, null.ok = TRUE)
+  checkmate::expect_character(drug_name_vec, null.ok = TRUE)
+  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
   checkmate::expect_character(colormap, null.ok = TRUE)
   checkmate::expect_flag(plot_averaged_flag)
   checkmate::expect_flag(plot_fit_flag)
   
-  grob_sa(
-    dt_metrics = gDRutils::convert_se_assay_to_dt(se, "Metrics"), 
-    dt_average = gDRutils::convert_se_assay_to_dt(se, "Averaged"), 
-    grouping = "rId",
-    normalization_type = normalization_type,
-    colormap = colormap,
-    plot_averaged_flag = plot_averaged_flag,
-    plot_fit_flag = plot_fit_flag
-  ) 
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  clid <- gDRutils::get_env_identifiers("cellline")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  gnumber <- gDRutils::get_env_identifiers("drug")
   
+  available_cellline <- unique(dt_metrics[[cellline_name]])
+  if (is.null(cellline_name_vec) || all(!cellline_name_vec %in% available_cellline)) {
+    cellline_name_vec <- available_cellline
+  } else if (!all(cellline_name_vec %in% available_cellline)) {
+    cellline_name_vec <- cellline_name_vec[cellline_name_vec  %in% available_cellline]
+  }  
+  
+  available_drugs <- unique(dt_metrics[[drug_name]])
+  if (is.null(drug_name_vec) || all(!drug_name_vec %in% available_drugs)) {
+    drug_name_vec  <- available_drugs
+  } else if (!all(drug_name_vec %in% available_drugs)) {
+    drug_name_vec <- drug_name_vec[drug_name_vec  %in% available_drugs]
+  }  
+  
+  plt_list <- list()
+  for (iC in cellline_name_vec) {
+    
+    # subset data
+    dt_metrics_subset <- dt_metrics[get(cellline_name) == iC & get(drug_name) %in% drug_name_vec]
+    dt_average_subset <- dt_average[get(cellline_name) == iC & get(drug_name) %in% drug_name_vec]
+    
+    plt_title <- sprintf("%s (%s)", iC, unique(dt_metrics[get(cellline_name) == iC, ][[clid]]))
+    
+    plt <- 
+      plot_dose_response_sa(dt_metrics = dt_metrics_subset, 
+                            dt_average = dt_average_subset, 
+                            grouping = drug_name,
+                            group_names = drug_name_vec,
+                            metric_growth = metric_growth,
+                            colormap = colormap,
+                            plot_averaged_flag = plot_averaged_flag,
+                            plot_fit_flag = plot_fit_flag)
+    
+    plt_list[[plt_title]] <- plt
+  }
+  
+  return(plt_list)
+}
+
+#' Plot drug response curves for single-agent data to check quality of data
+#'
+#' @param dt_metrics data.table representation of the data in \code{Metrics} assay
+#'    output from \code{gDRutils::convert_se_assay_to_dt(se, "Metrics")}
+#' @param dt_average data.table representation of the data in \code{Averaged} assay
+#'    output from \code{gDRutils::convert_se_assay_to_dt(se, "Averaged")}
+#' @param cl_name string cell line name to be plotted (Cell Line Name)
+#' @param d_name string vector with drug name to be plotted (Drug Name)
+#' @param metric_growth string with normalization_types to be selected
+#'                      one of: "GR" ("GRvalue") or "RV" ("RelativeViability")
+#' @param fit_source string source name for metrics
+#'
+#' @return plot with dose-response curves
+#'
+#' @keywords QC_plot
+#' @examples
+#' mae <- gDRutils::get_synthetic_data("small")
+#' se <- mae[[gDRutils::get_supported_experiments("sa")]]
+#'
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
+#' dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
+#' cl_name <- dt_metrics[["CellLineName"]][1]
+#' d_name <- dt_metrics[["DrugName"]][1]
+#' 
+#' plot_dose_response_sa_qc(dt_metrics = dt_metrics,
+#'                          dt_average = dt_average,
+#'                          cl_name = cl_name,
+#'                          d_name = d_name)
+#' 
+#' @export
+plot_dose_response_sa_qc <- function(dt_metrics, 
+                                     dt_average, 
+                                     cl_name, 
+                                     d_name,
+                                     metric_growth = "GR",
+                                     fit_source = "gDR") {
+  
+  checkmate::expect_data_table(dt_metrics)
+  checkmate::expect_data_table(dt_average)
+  checkmate::expect_string(cl_name)
+  checkmate::expect_string(d_name)
+  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
+  checkmate::assert_string(fit_source, null.ok = TRUE)
+  
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  clid <- gDRutils::get_env_identifiers("cellline")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  gnumber <- gDRutils::get_env_identifiers("drug")
+  conc <- gDRutils::get_env_identifiers("concentration")
+  
+  checkmate::expect_choice(cl_name, choices = dt_metrics[[cellline_name]])
+  checkmate::expect_choice(cl_name, choices = dt_average[[cellline_name]])
+  checkmate::expect_choice(d_name, choices = dt_metrics[[drug_name]])
+  checkmate::expect_choice(d_name, choices = dt_average[[drug_name]])
+  
+  # filter data for metric_growth
+  dt_metrics <- dt_metrics[normalization_type == metric_growth, ][
+    , .SD, .SDcols = c(drug_name, gnumber, cellline_name, clid, "x_inf", "x_0", "ec50", "h")]
+  dt_average <- dt_average[normalization_type == metric_growth, ][
+    , .SD, .SDcols = c(drug_name, gnumber, cellline_name, clid, conc, "x", "x_std")]
+  
+  selected_combination <- data.table::data.table(cellline_name = cl_name,
+                                                 drug_name = d_name)
+  data.table::setnames(selected_combination, c("cellline_name", "drug_name"), c(cellline_name, drug_name))
+  
+  # tab_plots
+  dt_average_plot <- dt_average[selected_combination, on = c(cellline_name, drug_name)]
+  dt_metrics_plot <- dt_metrics[selected_combination, on = c(cellline_name, drug_name)]
+  
+  if (NROW(dt_metrics_plot) > 0) {
+    min_conc <- min(dt_average_plot[get(conc) != 0][[conc]])
+    max_conc <- max(dt_average_plot[[conc]])
+    sampled_conc <- gDRplots::create_log_seq(min_conc, max_conc, 50) 
+    fitted_curve_sampled <- gDRutils::predict_efficacy_from_conc(sampled_conc,
+                                                                 dt_metrics_plot$x_inf,
+                                                                 dt_metrics_plot$x_0,
+                                                                 dt_metrics_plot$ec50,
+                                                                 dt_metrics_plot$h)
+    dt_reconstructed_fit <- data.table::data.table(
+      Concentration = sampled_conc,
+      x = fitted_curve_sampled
+    )
+    
+    # set min and max values for y 
+    ymin <- min(c(0, min(dt_average_plot$x)))
+    ymax <- max(c(1.2, max(dt_average_plot$x)))
+    
+    plt_title <- sprintf("%s (%s)", dt_metrics_plot[[drug_name]], dt_metrics_plot[[gnumber]])
+    
+    # plot
+    plt <- 
+      ggplot2::ggplot() +
+      ggplot2::geom_errorbar(
+        data = dt_average_plot, 
+        ggplot2::aes(x = get(conc), y = x,  ymin = x - x_std, ymax = x + x_std), 
+        width = 0.1, color = "#A9A9A9") + 
+      ggplot2::geom_line(
+        data = dt_average_plot, 
+        ggplot2::aes(x = get(conc), y = x), color = "black", linetype = "dashed") +
+      ggplot2::geom_line(
+        data = dt_reconstructed_fit, 
+        ggplot2::aes(x = get(conc), y = x), color = "red") +
+      ggplot2::geom_hline(yintercept = 0, color = "#A9A9A9") +
+      ggplot2::scale_x_continuous(trans = "log10") +
+      ggplot2::scale_y_continuous(lim = c(ymin, ymax)) +
+      ggplot2::xlab(bquote(.(conc) ~ "[" ~ mu * M ~ "]")) +
+      ggplot2::ylab(metric_growth) + 
+      ggplot2::ggtitle(plt_title) +
+      ggplot2::theme_bw() +
+      ggplot2::theme(panel.grid.minor = ggplot2::element_blank(),
+                     legend.position = "none")
+  } else {
+    txt_err <- sprintf(
+      "Dose response curve \nfor Drug Name: %s (%s) and CellLine: %s (%s) \n could not be calculated.",
+      dt_metrics_plot[[drug_name]], 
+      dt_metrics_plot[[gnumber]], 
+      dt_metrics_plot[[cellline_name]], 
+      dt_metrics_plot[[clid]])
+    plt <- 
+      ggplot2::ggplot() +
+      ggplot2::geom_text(ggplot2::aes(x = 0, y = 0, label = txt_err),
+                         color = "darkred", size = 5) + 
+      ggplot2::theme_void()
+  }
+  
+  return(plt)
+}
+
+#' Plot panel with drug response curves for single-agent data to check quality of data
+#'
+#' @inheritParams plot_dose_response_sa_qc
+#' @param d_names character vector with drug names to be plotted (Drug Name); 
+#'    if NULL - all available drugs will be plotted
+#'
+#' @return panle with plot with dose-response curves for selected cell line by drugs
+#'
+#' @keywords QC_plot
+#' @examples
+#' mae <- gDRutils::get_synthetic_data("small")
+#' se <- mae[[1]]
+#' 
+#' dt_metrics <- gDRutils::convert_se_assay_to_dt(se, "Metrics")
+#' dt_average <- gDRutils::convert_se_assay_to_dt(se, "Averaged")
+#' cl_name <- dt_metrics[["CellLineName"]][1]
+#' d_names <- unique(dt_metrics[["DrugName"]])[1:3]
+#' 
+#' plot_dose_response_sa_qc_panel(dt_metrics = dt_metrics,
+#'                                dt_average = dt_average,
+#'                                cl_name = cl_name)
+#' 
+#' plot_dose_response_sa_qc_panel(dt_metrics = dt_metrics,
+#'                                dt_average = dt_average,
+#'                                cl_name = cl_name,
+#'                                d_names = d_names)
+#' 
+#' @export
+plot_dose_response_sa_qc_panel <- function(dt_metrics, 
+                                           dt_average, 
+                                           cl_name, 
+                                           d_names = NULL,
+                                           metric_growth = "GR",
+                                           fit_source = "gDR") {
+  
+  checkmate::expect_data_table(dt_metrics)
+  checkmate::expect_data_table(dt_average)
+  checkmate::expect_string(cl_name)
+  checkmate::expect_character(d_names, null.ok = TRUE)
+  checkmate::expect_choice(metric_growth, choices = c("GR", "RV"))
+  checkmate::assert_string(fit_source, null.ok = TRUE)
+  
+  cellline_name <- gDRutils::get_env_identifiers("cellline_name")
+  clid <- gDRutils::get_env_identifiers("cellline")
+  drug_name <- gDRutils::get_env_identifiers("drug_name")
+  gnumber <- gDRutils::get_env_identifiers("drug")
+  conc <- gDRutils::get_env_identifiers("concentration")
+  
+  checkmate::expect_choice(cl_name, choices = dt_metrics[[cellline_name]])
+  checkmate::expect_choice(cl_name, choices = dt_average[[cellline_name]])
+  
+  available_drugs <- unique(dt_metrics[[drug_name]])
+  if (is.null(d_names) || all(!d_names %in% available_drugs)) {
+    d_names  <- available_drugs
+  } else if (!all(d_names %in% available_drugs)) {
+    d_names <- drug_name[drug_name  %in% available_drugs]
+  } 
+  
+  ls_drug <- list(d_name = d_names)
+  
+  # list of plots for each drug
+  ls_plt <- purrr::pmap(ls_drug, 
+                        gDRplots::plot_dose_response_sa_qc,
+                        dt_metrics = dt_metrics,
+                        dt_average = dt_average,
+                        cl_name = cl_name,
+                        metric_growth = metric_growth,
+                        fit_source = fit_source)
+  
+  # panel title
+  cl_clid <- unique(dt_metrics[get(cellline_name) == cl_name, ][[clid]]) 
+  panel_title <- sprintf("%s (%s)", cl_name, cl_clid)
+  
+  # final panel
+  ggpubr::annotate_figure(ggpubr::ggarrange(plotlist = ls_plt),
+                          top = panel_title)
 }