+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.7 - 2024-06-21
+* add plot to quality check of dose response
+
 ## gDRplots 0.0.6 - 2024-06-19
 * update plot function for `processingReport` package